Add placeholder routes
