+++ app/routes/elements.tsx
@@ -1,0 +1,13 @@
+import type { Route } from './+types/home';
+import { TextVariant } from '~/elements/text';
+
+export function meta({}: Route.MetaArgs) {
+	return [
+		{ title: 'New React Router App' },
+		{ name: 'description', content: 'Welcome to React Router!' },
+	];
+}
+
+export default function Elements() {
+	return <ui-text data-variant={TextVariant.XL}>Elements</ui-text>;
+}
