+++ package.json
@@ -16,6 +16,7 @@
 		"typecheck": "react-router typegen && tsc"
 	},
 	"dependencies": {
+		"@heroicons/react": "^2.2.0",
 		"@lit/react": "^1.0.8",
 		"@react-router/node": "^7.7.1",
 		"@react-router/serve": "^7.7.1",
